Basic form logic

+++ src/Controllers/Table.java
@@ -2,29 +2,36 @@ package Controllers;
 
 import Models.Item;
 import javafx.beans.property.SimpleIntegerProperty;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
+import java.util.function.Consumer;
 
-public abstract class Table extends Base implements Initializable {
+public abstract class Table<T extends Item> extends Base implements Initializable {
     @FXML
-    protected TableView<Item> tableView;
+    protected TableView<T> tableView;
 
     protected abstract void addColumns();
 
     protected abstract void populateData();
 
-    protected Form formController;
+    protected Form<T> formController;
+    protected Optional<Integer> userId;
+    private FormFactory formFactory;
+
+    public Table(Optional<Integer> userId, FormFactory formFactory) {
+        this.userId = userId;
+        this.formFactory = formFactory;
+    }
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-        final TableColumn<Item, Integer> idColumn = new TableColumn("ID");
+        final TableColumn<T, Integer> idColumn = new TableColumn("ID");
         idColumn.setCellValueFactory(param -> new SimpleIntegerProperty(param.getValue().getId()).asObject());
         tableView.getColumns().add(idColumn);
         addColumns();
@@ -32,50 +39,69 @@ public abstract class Table extends Base implements Initializable {
         tableView.refresh();
     }
 
-    private Item openForm(Item item, Form.Mode mode) {
-        formController.open(item, mode);
-        return item;
+    private void openForm(FormFactory.Type type, T item, Form.Mode mode, Consumer<T> callback) {
+        formController = formFactory.getInstance(type);
+        formController.open(item, mode, callback);
     }
 
-    protected abstract void addToDatabase(Item item);
+    private void finalizeAction() {
+        tableView.refresh();
+        formController = null;
+    }
 
-    protected abstract Item getNewItem();
+    protected abstract void addToDatabase(T item);
+
+    protected abstract T getNewItem();
 
     @FXML
     private void addItem() {
-        System.out.println("add item called");
-        Item newItem = openForm(getNewItem(), Form.Mode.Create);
-        addToDatabase(newItem);
-        tableView.refresh();
+        if (formController == null) {
+            System.out.println("add item called");
+            openForm(FormFactory.Type.Create, getNewItem(), Form.Mode.Create, (newItem) -> {
+                if (newItem != null) addToDatabase(newItem);
+                finalizeAction();
+            });
+        }
     }
 
-    private Item getSelectedItem() {
+    private T getSelectedItem() {
         return tableView.getSelectionModel().getSelectedItem();
     }
 
     @FXML
     private void viewItem() {
-        System.out.println("view item called");
-        openForm(getSelectedItem(), Form.Mode.Read);
+        T selected = getSelectedItem();
+        System.out.println(selected);
+        if (selected != null && formController == null) {
+            System.out.println("view item called");
+            openForm(FormFactory.Type.Read, selected, Form.Mode.Read, (item) -> finalizeAction());
+        }
     }
 
-    protected abstract void updateInDatabase(Item item);
+    protected abstract void updateInDatabase(T item);
 
     @FXML
     private void editItem() {
-        System.out.println("edit item called");
-        updateInDatabase(openForm(getSelectedItem(), Form.Mode.Update));
-        tableView.refresh();
+        T selected = getSelectedItem();
+        if (selected != null && formController == null) {
+            System.out.println("edit item called");
+            openForm(FormFactory.Type.Update, selected, Form.Mode.Update, (updatedItem) -> {
+                if (updatedItem != null) updateInDatabase(updatedItem);
+                finalizeAction();
+            });
+        }
     }
 
-    protected abstract void deleteFromDatabase(Item item);
+    protected abstract void deleteFromDatabase(T item);
 
     @FXML
     private void deleteItem() {
         System.out.println("delete item called");
-        Item itemToDelete = getSelectedItem();
-        deleteFromDatabase(itemToDelete);
-        tableView.getItems().remove(itemToDelete);
-        tableView.refresh();
+        T itemToDelete = getSelectedItem();
+        if (itemToDelete != null) {
+            deleteFromDatabase(itemToDelete);
+            tableView.getItems().remove(itemToDelete);
+            tableView.refresh();
+        }
     }
 }
